restore climbing cost. had to put stuff back to float

+++ TLBGridPoints/Data/Scripts/ShipPoints/BuildingBlockPoints.cs
@@ -56,8 +56,8 @@ namespace TLB.ShareTrack
 
             string blockDisplayName = blockInfo.BlockName;
 
-            float thisClimbingCostMult = 0;
-            AllGridsList.ClimbingCostRename(ref blockDisplayName, ref thisClimbingCostMult);
+            double thisClimbingCostMult = 0;
+            AllGridsList.ClimbingCostRename(ref blockDisplayName, ref thisClimbingCostMult);  // Use double instead of float.
 
             _pointsMessage.Message.Clear();
             _pointsMessage.Message.Append($"{blockDisplayName}:\n{blockPoints}bp");
